Move nirspec parser test
