+++ jdaviz/configs/mosviz/tests/test_data_loading.py
@@ -186,89 +186,6 @@ def test_load_single_image_multi_spec(mosviz_helper, mos_image, spectrum1d, mos_
     assert image_viewer.label_mouseover.world_dec_deg == ''
 
 
-@pytest.mark.filterwarnings('ignore')
-@pytest.mark.remote_data
-def test_nirspec_loader(mosviz_helper, tmpdir):
-    '''
-    Tests loading our default MosvizExample notebook data
-    Also tests IntraRow linking
-    '''
-
-    test_data = 'https://stsci.box.com/shared/static/ovyxi5eund92yoadvv01mynwt8t5n7jv.zip'
-    fn = download_file(test_data, cache=True, timeout=30)
-    with ZipFile(fn, 'r') as sample_data_zip:
-        sample_data_zip.extractall(tmpdir)
-
-    level3_path = (pathlib.Path(tmpdir) / 'mosviz_nirspec_data_0.3' / 'level3')
-
-    data_dir = level3_path
-
-    mosviz_helper.load_data(directory=data_dir, instrument='nirspec')
-
-    assert len(mosviz_helper.app.data_collection) == 16
-
-    dc_0 = mosviz_helper.app.data_collection[0]
-    assert dc_0.label == "MOS Table"
-    assert len(dc_0.meta) == 0
-
-    dc_5 = mosviz_helper.app.data_collection[5]
-    assert dc_5.label == "Image 4"
-    assert PRIHDR_KEY not in dc_5.meta
-    assert dc_5.meta['WCSAXES'] == 2
-
-    dc_10 = mosviz_helper.app.data_collection[10]
-    assert dc_10.label == "1D Spectrum 4"
-    assert PRIHDR_KEY not in dc_10.meta
-    assert 'header' not in dc_10.meta
-    assert dc_10.meta['TARGNAME'] == 'FOO'
-
-    dc_15 = mosviz_helper.app.data_collection[15]
-    assert dc_15.label == "2D Spectrum 4"
-    assert PRIHDR_KEY not in dc_15.meta
-    assert 'header' not in dc_15.meta
-    assert dc_15.meta['SOURCEID'] == 2315
-
-    # Test IntraRow linking:
-    # Attempts to add the spectrum of another row into the current row's viewers
-    # Currently, intrarow linking is disabled. Attemps to load another spectrum into
-    # the current spectrum viewer should result in an error
-
-    # Check to make sure our test case isn't from the same row
-    table = mosviz_helper.app.get_viewer('table-viewer')
-    table.select_row(0)
-    data_label = "1D Spectrum 4"
-    assert mosviz_helper.app.data_collection[data_label].meta['mosviz_row'] != table.current_row
-
-    with pytest.raises(NotImplementedError, match='Intra-row plotting not supported'):
-        mosviz_helper.app.add_data_to_viewer(viewer_reference='spectrum-viewer',
-                                             data_label=data_label)
-
-
-@pytest.mark.remote_data
-def test_nirspec_fallback(mosviz_helper, tmpdir):
-    '''
-    When no instrument is provided, mosviz.load_data is expected to fallback to the nirspec loader.
-    Naturally, the nirspec dataset should then work without any instrument keyword
-    '''
-
-    test_data = 'https://stsci.box.com/shared/static/ovyxi5eund92yoadvv01mynwt8t5n7jv.zip'
-    fn = download_file(test_data, cache=True, timeout=30)
-    with ZipFile(fn, 'r') as sample_data_zip:
-        sample_data_zip.extractall(tmpdir)
-
-    level3_path = (pathlib.Path(tmpdir) / 'mosviz_nirspec_data_0.3' / 'level3')
-
-    data_dir = level3_path
-    with pytest.warns(UserWarning, match="Ambiguous MOS Instrument"):
-        mosviz_helper.load_data(directory=data_dir)
-
-    assert len(mosviz_helper.app.data_collection) == 16
-    assert "MOS Table" in mosviz_helper.app.data_collection
-    assert "Image 4" in mosviz_helper.app.data_collection
-    assert "1D Spectrum 4" in mosviz_helper.app.data_collection
-    assert "2D Spectrum 4" in mosviz_helper.app.data_collection
-
-
 def test_zip_error(mosviz_helper, tmp_path):
     '''
     Zipfiles are explicitly and intentionally not supported. This test confirms a TypeError is
